Coding Etiquette: Adjust Model Names
Hi Isaac. I am not super well versed in the goal of this project, but it seems you create and run models in two different files. I would avoid using the same names for models you are running in different files; although separated and in a logical order this avoids confusion if different names are used.