Coding Etiquette: Naming of Dataframes
Hi Isaac! I would name your dataframes and tables something more insightful to a user jumping on the project instead of "eie". If this represents something meaningful it would be a great opportunity to share that in a vignette or the readme file if you would prefer to keep it.